Change Flex-Properties to Account for Width
Some elements were not shrinking properly when the window was resized.
This is because these elements were using padding to squeeze them
together instead of a more organic property of flexbox such as
space-around or space-evenly.  I changed the width of the top-header
logo to equal the width of the top-content so that the both sections
would have the same amount of space.